+++ src/index.tsx
@@ -3,9 +3,8 @@ import { HashRouter as Router } from 'react-router-dom';
 import { CssBaseline } from '@mui/material';
 import { Provider } from 'react-redux';
 
-import { store } from './redux/store';
-import App from './App';
-import './index.css';
+import { store } from './shared/store';
+import App from './app';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
