+++ src/widgets/footer/assets/index.d.ts
