+++ src/widgets/footer/index.tsx
@@ -1,7 +1,7 @@
 import { memo } from 'react';
 import { Toolbar, Typography, Link, SvgIcon } from '@mui/material';
 
-import { ReactComponent as RSS } from '../assets/logo_RSS.svg';
+import { ReactComponent as RSS } from './assets/logo_RSS.svg';
 
 const Footer = () => (
   <Toolbar sx={{ bgcolor: '#000', paddingY: 1 }}>
