+++ src/widgets/layout/index.tsx
@@ -1,8 +1,8 @@
 import { Fragment } from 'react';
 import { Outlet } from 'react-router-dom';
 
-import Footer from './Footer';
-import Header from './Header';
+import Footer from '../footer';
+import Header from '../header';
 
 const Layout = () => {
   return (
